Clean up and rename no_shunt_ivr to reduced_ivr
